feat: DatePicker로 저장한 날짜값을 요구된 형태로 변환하여 저장할 수 있도록 구현

+++ ProjectManager/ProjectManager/Source/Model/Model/ProjectData.swift
@@ -8,5 +8,5 @@
 struct ProjectData {
     let title: String
     let body: String
-    let deadline: Int
+    let deadline: Double
 }
